pyproject.toml
[tool.poetry]
name = "django-rest-framework-jwk"
version = "0.1.0"
description = "Easy JSON Web Keys (JWK) for your Django project"
authors = ["Gustavo Maronato <gustavo@maronato.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
cryptography = "^3.3.1"
Django = "^2.2"
djangorestframework = "^3.10"
jwcrypto = "^0.8"

[tool.poetry.dev-dependencies]
pytest-django = "4.1.0"
pytest = "6.2.1"
pytest-cov = "2.10.1"
flake8 = "3.8.4"
mkdocs = "1.1.2"
tox = "^3.20.1"
mypy = "^0.790"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
